Replace Redirected URLs
This follows the redirects of any packages added and checks to make sure the end URL is not different to the one provided.

If it is different, we override the one they provided and perform all the other checks against that new URL.

This also includes a change to return a failure exit code should any of the "auto fix" errors be raised. This will prompt us on CI should a user push up before running this script and commiting the changes.

This fixes #522 (SPI-Server Repo)