[autopr] allow validate failure, remove unnecessary dupe fix
The validate step (correctly) returns a failure exit code when it makes changes. We don't want this to fail CI as we're relying on it making those changes for us for the auto PR. What it outputs should be valid JSON which can be pushed to a PR.

Also removed the dupe/sort step in add_package as this is done by the validate step so no point duplicating effort.