[autopr] add CI step to automatically create PR to remove chosen packages
Mostly duplicate of the "Add Package" route with some minor tweaks especially with the JQ line.

We will convert SPI links to GitHub links, in case a user chooses to do that. We're resilient to URLs with/without the ".git" at the end. We lowercase both sides of the search because our packages.json is not case sensitive.